feat: indexing and querying (#5)
* update list index tasks

* fix: make progress floaty

* add main view

* fix: index task failure on db constraint

* some ev work

* todo: some good processing

* wip: figuring out hnsw_rs's weird io interface

* refactor: give up on saving vector db

Doing so would take way too long to feasibly accomplish within the next
few hours, so I'm going to forego this process. Ideally, you would be
able to save your store alongside the sqlite db. happens.

* remove dead code

+++ src/models/_entities/index_tasks.rs
@@ -3,7 +3,7 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
 #[sea_orm(table_name = "index_tasks")]
 pub struct Model {
     pub created_at: DateTimeWithTimeZone,
@@ -11,7 +11,8 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub path: String,
-    pub progress: i32,
+    #[sea_orm(column_type = "Float")]
+    pub progress: f32,
     pub queue: String,
 }
